refactor: move decoder to dedicated file
