+++ subgraph/src/utils/contract-calls.ts
@@ -1,7 +1,6 @@
-import { Address, BigInt, Bytes, ethereum, log } from "@graphprotocol/graph-ts";
+import { Address, BigInt, Bytes, log } from "@graphprotocol/graph-ts";
 import { ServiceProviderRegistry } from "../../generated/ServiceProviderRegistry/ServiceProviderRegistry";
-import { PDP_OFFERING_DEF } from "./constants";
-import { PDPOffering, ServiceProviderInfo } from "./types";
+import { ServiceProviderInfo } from "./types";
 import { PDPVerifier } from "../../generated/PDPVerifier/PDPVerifier";
 
 export function getServiceProviderInfo(registryAddress: Address, providerId: BigInt): ServiceProviderInfo {
@@ -50,26 +49,4 @@ export function getPieceCidData(verifierAddress: Address, setId: BigInt, pieceId
   }
 
   return pieceCidTry.value.data;
-}
-
-export function decodePDPOfferingData(data: Bytes): PDPOffering {
-  const decoded = ethereum.decode(PDP_OFFERING_DEF, data);
-
-  if (!decoded) {
-    log.warning("[decodePDPOfferingData]: failed to decode data: {}", [data.toHexString()]);
-    return PDPOffering.empty();
-  }
-
-  const decodedTuple = decoded.toTuple();
-  return new PDPOffering(
-    decodedTuple[0].toString(),
-    decodedTuple[1].toBigInt(),
-    decodedTuple[2].toBigInt(),
-    decodedTuple[3].toBoolean(),
-    decodedTuple[4].toBoolean(),
-    decodedTuple[5].toBigInt(),
-    decodedTuple[6].toBigInt(),
-    decodedTuple[7].toString(),
-    decodedTuple[8].toAddress(),
-  );
 }
